revert #78
The original code was right. It's a queue, so q->fi points the last element
in the queue, unless the queue is empty. If the queue is not empty, the old
last element should point new last (e), and q->fi should move to e. If the
queue is empty q->fi is NULL, and should be updated to e.