add generic function
a generic function is a function-like object that mimics a call under
the context at the point it is obtained.

to get a generic function: &func
to call a function through a generic function: gf.(args) or gf(args)
note that the latter form may be overridden by the first argument
namespace. To ensure the generic function to be called, use the former
style.